+++ gluecodium/src/main/java/com/here/gluecodium/generator/common/CommentsProcessor.kt
@@ -23,11 +23,13 @@ import com.here.gluecodium.common.LimeLogger
 import com.vladsch.flexmark.ast.AutoLink
 import com.vladsch.flexmark.ast.Code
 import com.vladsch.flexmark.ast.LinkRef
+import com.vladsch.flexmark.ext.tables.TablesExtension
 import com.vladsch.flexmark.parser.Parser
 import com.vladsch.flexmark.util.ast.IRender
 import com.vladsch.flexmark.util.ast.NodeVisitor
 import com.vladsch.flexmark.util.ast.VisitHandler
-import com.vladsch.flexmark.util.data.DataSet
+import com.vladsch.flexmark.util.data.DataHolder
+import com.vladsch.flexmark.util.data.MutableDataSet
 import com.vladsch.flexmark.util.sequence.BasedSequenceImpl
 
 /**
@@ -39,7 +41,7 @@ abstract class CommentsProcessor(private val renderer: IRender, private val werr
         get() = werror && hasErrorFlag
     private var hasErrorFlag = false
 
-    private val parser = Parser.builder(DataSet()).build()
+    private val parser = Parser.builder(flexmarkOptions).build()
     private val logFunction: LimeLogger.(String, String) -> Unit =
         if (werror) { elementName: String, message: String ->
             this.error(elementName, message)
@@ -105,5 +107,8 @@ abstract class CommentsProcessor(private val renderer: IRender, private val werr
 
     companion object {
         private const val standardNullReference = "null"
+        internal val flexmarkOptions: DataHolder = MutableDataSet()
+            .set(Parser.EXTENSIONS, listOf(TablesExtension.create()))
+            .toImmutable()
     }
 }